webapi/webapi-webrtc-w3c-tests/webrtc/RTCIdentityEvent_assertion_basic.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Liu,Yun <yunx.liu@intel.com>

-->

<meta charset="utf-8">
<title>WebRTC Test: RTCIdentityEvent - assertion basic</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://dev.w3.org/2011/webrtc/editor/webrtc.html#rtcidentityevent-type">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./support/support.js"></script>
<div id="log"></div>
<script>

var t = async_test("Check if the attribute assertion of RTCIdentityEvent exists and type of string");
setup({timeout: 20000});

var pc1_offer;
var pc2_answer;
var pc1;
if (typeof RTCPeerConnection != "undefined") {
  pc1 = new RTCPeerConnection(null, null);
} else {
  pc1 = new webkitRTCPeerConnection(null, null);
}
var pc2;

function requestSucceeded1 (offer) {
  pc1_offer = offer;
  pc1.setLocalDescription(offer, requestSucceeded2, requestFailed2);
}

function requestFailed1 () {
  t.step(function () {
    assert_unreached("requestFailed1 was called");
  });
  t.done();
}

function requestSucceeded2 () {
  if (typeof RTCPeerConnection != "undefined") {
    pc2 = new RTCPeerConnection(null, null);
  } else {
    pc2 = new webkitRTCPeerConnection(null, null);
  }
  pc2.setRemoteDescription(pc1_offer, requestSucceeded3, requestFailed3);
}

function requestFailed2 () {
  t.step(function () {
    assert_unreached("requestFailed2 was called");
  });
  t.done();
}

function requestSucceeded3 () {
  pc2.createAnswer(requestSucceeded4, requestFailed4);
}

function requestFailed3 () {
  t.step(function () {
    assert_unreached("requestFailed3 was called");
  });
  t.done();
}

function requestSucceeded4 (answer) {
  pc2_answer = answer;
  pc2.setLocalDescription(answer, requestSucceeded5, requestFailed5);
}

function requestFailed4 () {
  t.step(function () {
    assert_unreached("requestFailed4 was called");
  });
  t.done();
}

function requestSucceeded5 () {
  pc1.setRemoteDescription(pc2_answer, function () {}, function () {
    t.step(function () {
      assert_unreached("setRemoteDescription request failed");
    });
    t.done();
  });
}
function requestFailed5 () {
  t.step(function () {
    assert_unreached("requestFailed5 was called");
  });
  t.done();
}

t.step(function () {
  pc1.createDataChannel("This is pc1");
  pc1.onidentityresult = t.step_func(function (event) {
    assert_true("assertion" in event, "assertion attribute in RTCIdentityEvent");
    assert_equals(Object.prototype.toString.call(event.assertion), "[object String]", "assertion attribute of type");
    t.done();
  });
  pc1.createOffer(requestSucceeded1, requestFailed1);
});

</script>
